Add check for invalid userAddress
Thought I'd take a crack at preventing this bug from happening again. Maybe it would make more sense to do this in the existing checks that happen in `checkChainId` but this seems like it should at cover this edge case. Full disclosure: I haven't written any Vue or web3 code before 😅  but I tried to follow the conventions I found in the code that was here.